+++ projects/rocprofiler-sdk/tests/rocprofv3/counter-collection/extended-options/pytest.ini
@@ -1,5 +1,5 @@
 
 [pytest]
 addopts = --durations=20 -rA -s
-testpaths = test_extended_options.py
+testpaths = validate.py
 pythonpath = @ROCPROFILER_SDK_TESTS_BINARY_DIR@/pytest-packages
